workflow: '24.05' -> 'stable'
The expected flow in the build script was not followed. I'm not sure if
this was a mistake or if the intentions changed later on, but
`./scripts/build_and_parse_hm_options.rb` checked for `HM_RELEASE` to be
'stable' to then use the branch referenced in `./config.yaml` under the
`release_current_stable` key. Based on the instructions for updating the
stable branch reference, this is what is intended. The workflow however
doesn't have this mapped correctly, rather `release-24.05` is being
built explicitly, rather than "stable".

Although `release-24.11` was added, its docs were never actually built.
This has been corrected here to follow what I interpret as the intended
workflow. If it is preferred to simply update the workflow manually to
point to the new version each time, then this can be done.